+++ frontend/src/components/ui/carousels/GuideCarousel/Mobile/index.tsx
@@ -1,0 +1,51 @@
+'use client'
+
+import type { EmblaOptionsType } from 'embla-carousel'
+import Autoplay from 'embla-carousel-autoplay'
+import { carouselData } from '~/lib/data/carouselData'
+import LogoImage from '../../../LogoImage'
+import {
+  Carousel,
+  CarouselContent,
+  CarouselItem,
+} from '../../Carousel'
+import GuideCard from './GuideCard'
+
+const carouselOptions: Partial<EmblaOptionsType> = {
+  align: 'start',
+  loop: true,
+}
+
+const carouselPlugins = [
+  Autoplay({
+    delay: 4000,
+    stopOnInteraction: false,
+  }),
+]
+
+export default function MobileGuideCarousel() {
+  return (
+    <>
+      <h2 className="mb-7 text-center text-lg text-secondary-foreground sm:text-2xl md:mb-12">
+        <span className="inline-block align-text-bottom">
+          <LogoImage
+            width={170}
+            height={30}
+            className="w-44"
+          />
+        </span>
+        <span className="inline-block pl-0.5">の使い方</span>
+      </h2>
+
+      <Carousel opts={carouselOptions} plugins={carouselPlugins}>
+        <CarouselContent>
+          {carouselData.map(data => (
+            <CarouselItem key={data.id}>
+              <GuideCard data={data} />
+            </CarouselItem>
+          ))}
+        </CarouselContent>
+      </Carousel>
+    </>
+  )
+}
